api/requirements.txt
azure-core==1.32.0
azure-identity==1.20.0
azure-storage-blob==12.24.1
blinker==1.8.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.1
et-xmlfile==2.0.0
flask==3.0.3
Flask-Cors==5.0.0
flask-swagger-ui==4.11.1
idna==3.10
importlib-metadata==8.5.0
isodate==0.7.2
itsdangerous==2.2.0
jinja2==3.1.5
MarkupSafe==2.1.5
msal==1.31.1
msal-extensions==1.2.0
numpy==1.24.4
openpyxl==3.1.5
pandas==2.0.3
portalocker==2.10.1
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
six==1.17.0
typing-extensions==4.12.2
tzdata==2025.1
urllib3==2.2.3
werkzeug==3.0.6
zipp==3.20.2
